Bug Fix - Admin Login & Logout
The issue was with how the session records were kept. They were stored in a single table with a foreign key to userId. So, another table named admin_sessions had to be created. The logic also updated accordingly.

code to run on the database:
CREATE TABLE gamage_db.admin_sessions (
  SessionId INT AUTO_INCREMENT PRIMARY KEY,
  Id INT NOT NULL,
  token VARCHAR(256) NOT NULL,
  createdAt DATETIME NOT NULL,
  status VARCHAR(20) NOT NULL,
  role VARCHAR(20) NOT NULL,
  FOREIGN KEY (Id) REFERENCES admin(adminId) ON DELETE CASCADE
);